cloud_relay/requirements.txt
fastapi>=0.75.0
uvicorn>=0.17.0
pyjwt>=2.3.0
python-multipart>=0.0.5
websockets>=10.2
pydantic>=1.9.0
python-dotenv>=0.20.0
cryptography>=36.0.0
prometheus-client>=0.14.0
httpx>=0.22.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
python-jose>=3.3.0 